+++ packages/extension-base/src/stores/Metadata.ts
@@ -8,6 +8,10 @@ import BaseStore from './Base';
 
 export default class MetadataStore extends BaseStore<MetadataDef> {
   constructor () {
-    super(`${EXTENSION_PREFIX}metadata`);
+    super(
+      EXTENSION_PREFIX && EXTENSION_PREFIX !== 'polkadot{.js}'
+        ? `${EXTENSION_PREFIX}metadata`
+        : 'metadata'
+    );
   }
 }