+++ packages/extension/webpack.shared.cjs
@@ -3,12 +3,15 @@
 
 const path = require('path');
 const webpack = require('webpack');
-
 const CopyPlugin = require('copy-webpack-plugin');
 const ManifestPlugin = require('webpack-extension-manifest-plugin');
 
+const { blake2AsHex } = require('@polkadot/util-crypto');
+
 const pkgJson = require('./package.json');
 const manifest = require('./manifest.json');
+
+const EXT_NAME = manifest.short_name;
 
 const packages = [
   'extension',
@@ -70,7 +73,9 @@ module.exports = (entry, alias = {}) => ({
     }),
     new webpack.DefinePlugin({
       'process.env': {
-        NODE_ENV: JSON.stringify('production')
+        EXTENSION_PREFIX: JSON.stringify(process.env.EXTENSION_PREFIX || EXT_NAME),
+        NODE_ENV: JSON.stringify('production'),
+        PORT_PREFIX: JSON.stringify(blake2AsHex(JSON.stringify(manifest), 64))
       }
     }),
     new CopyPlugin({ patterns: [{ from: 'public' }] }),